seperate fsdr utils and clean
